starting point
As I witten in the README I have struggled with making the social links vertical on my desktop and this is as far as I have come. At the time I hadent looked at the plagiarism information and this is therefor my first commit to this project, I have started this a few times now but I never got satisfid so now I have dona a wireframe. I will change the image and have borders on the side. the image will be a triangle in the middle, look at the wireframe